Added cmake-stamp

+++ library_builders.sh
@@ -200,20 +200,29 @@ function get_modern_cmake {
     # Install cmake >= 2.8
     local cmake=cmake
     if [ -n "$IS_MACOS" ]; then
-        brew install cmake > /dev/null
+        if [ ! -e cmake-stamp ]; then
+            brew install cmake > /dev/null
+        fi
     elif [ -n "$IS_ALPINE" ]; then
-        apk add cmake > /dev/null
+        if [ ! -e cmake-stamp ]; then
+            apk add cmake > /dev/null
+        fi
     elif [[ $MB_ML_VER == "_2_24" ]]; then
         # debian:9 based distro
-        apt-get install -y cmake
+        if [ ! -e cmake-stamp ]; then
+            apt-get install -y cmake
+        fi
     else
         if [ "`yum search cmake | grep ^cmake28\.`" ]; then
             cmake=cmake28
         fi
         # centos based distro
-        yum_install $cmake > /dev/null
+        if [ ! -e cmake-stamp ]; then
+            yum_install $cmake > /dev/null
+        fi
     fi
     echo $cmake
+    touch cmake-stamp
 }
 
 function get_cmake {
